requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
PyMySQL==1.1.0
pandas==2.1.4

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM integrations
openai==1.3.7
requests==2.31.0

# Local LLM support
gpt4all==2.0.2

# SQL parsing and validation
sqlparse==0.4.4
sqlglot==19.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
